BL2: Backport some changes from mcuboot PR#2305
This patch backports updates from MCUboot PR#2305 that were missed in the
existing TF-M integration patches.

Key changes:
- Removed outdated hardware rollback protection handling from
  `bootutil/CMakeLists.txt` as mcuboot config are handled via header.
- Cleaned up unused definitions in `bootutil/include/bootutil/image.h`.
- Extended `bootutil/include/bootutil/sign_key.h` to include additional
  key types and helper macros.
- Minor corrections in `bootutil_find_key.c` and image validation paths.
- Synced image hashing and validation function naming with MCUboot.
- Added release note: `multi-signature-support.md`.
- Updated `mcuboot_config.template.h` with new signature configuration
  macros.
- Improved Python `imgtool/image.py` to support multi-signature options.

For further details, please refer to comments on the mcuboot PR.

+++ boot/bootutil/src/bootutil_priv.h
@@ -291,7 +291,7 @@ fih_ret bootutil_verify_sig(uint8_t *hash, uint32_t hlen, uint8_t *sig,
  * against provided signature.
  */
 fih_ret bootutil_verify_img(uint8_t *img, uint32_t size,
-                            uint8_t *sig, size_t slen, uint8_t key_id);
+                            uint8_t *sig, size_t slen, uint32_t key_id);
 
 fih_ret boot_fih_memequal(const void *s1, const void *s2, size_t n);
